+++ credmark/cmf/types/data/fungible_token_data_bsc.py
@@ -1,0 +1,39 @@
+BSC_TOKENS = {
+    'ETH': {
+        'symbol': 'ETH',
+        'decimals': 18,
+        'name': 'Ethereum Token',
+        'address': '0x2170Ed0880ac9A755fd29B2688956BD959F933F8',
+    },
+    'USDT': {
+        'symbol': 'USDT',
+        'decimals': 18,
+        'name': 'Tether USD',
+        'address': '0x55d398326f99059fF775485246999027B3197955',
+    },
+    'BUSD': {
+        'symbol': 'BUSD',
+        'decimals': 18,
+        'name': 'BUSD Token',
+        'address': '0xe9e7CEA3DedcA5984780Bafc599bD69ADd087D56',
+    },
+    'USDC': {
+        'symbol': 'USDC',
+        'decimals': 18,
+        'name': 'USD Coin',
+        'address': '0x8AC76a51cc950d9822D68b83fE1Ad97B32Cd580d',
+    },
+    'WBNB': {
+        'symbol': 'WBNB',
+        'decimals': 18,
+        'name': 'Wrapped BNB',
+        'address': '0xbb4CdB9CBd36B01bD1cBaEBF2De08d9173bc095c',
+        "is_native_token": True,
+    },
+    'BTCB': {
+        'symbol': 'BTCB',
+        'decimals': 18,
+        'name': 'BTCB Token',
+        'address': '0x7130d2A12B9BCbFAe4f2634d864A1Ee1Ce3Ead9c',
+    },
+}